fix: remove use_manual_address_prefixes variable  (#510)
BREAKING CHANGE:

Removed the input variable `use_manual_address_prefixes`.  The value of this variable is now computed by the module based on the `address_prefixes` input variable. If `address_prefixes` variable is set, then the modules determine that manual address prefix are used.

Note: in the event that `use_manual_address_prefix=false` (the default value) AND `address_prefixes` was specified, the VPC and everything deployed into it will be destroyed and it will be recreated when moving up to this version. To avoid this - BEFORE upgrading to this version of the module - set `use_manual_address_prefix` to true, and apply the change (`terraform apply`). You are now ready to upgrade to the new version without ending up in a VPC deletion.